requirements_updated.txt
Django==4.2.26
pymysql==1.1.0
python-decouple==3.8
Pillow==12.1.0
gunicorn==23.0.0
PyJWT==2.11.0
pyotp==2.9.0
qrcode==7.4.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
cryptography==42.0.0
django-allauth==0.57.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
requests==2.31.0
argon2-cffi==23.1.0
pycryptodome==3.20.0
psutil==5.9.8
